+++ libs/platform/src/lib/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.43.0-rc.11](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.10...v0.43.0-rc.11) (2023-07-13)
+
+
+### Bug Fixes
+
+* **platform:** remove aria-labelledby from platform table header cells ([#10157](https://github.com/SAP/fundamental-ngx/issues/10157)) ([41877a3](https://github.com/SAP/fundamental-ngx/commit/41877a3ff4541623d32f0913c261961614d9080e))
+
+
+
+
+
+# [0.43.0-rc.10](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.9...v0.43.0-rc.10) (2023-07-13)
+
+
+### Bug Fixes
+
+* **platform:** table console error ([#10163](https://github.com/SAP/fundamental-ngx/issues/10163)) ([3228343](https://github.com/SAP/fundamental-ngx/commit/3228343906f527c1f2b0455ba582af6bc6f7d91f))
+
+
+
+
+
+# [0.43.0-rc.9](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.8...v0.43.0-rc.9) (2023-07-13)
+
+
+### Bug Fixes
+
+* **docs,core,platform:** various fixes from deffect hunt ([#10168](https://github.com/SAP/fundamental-ngx/issues/10168)) ([00b858b](https://github.com/SAP/fundamental-ngx/commit/00b858bcfb1a1e6963c36a7d768fe9dcf100a135))
+
+
+
+
+
+# [0.43.0-rc.8](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.7...v0.43.0-rc.8) (2023-07-13)
+
+
+### Bug Fixes
+
+* **platform:** fixed table sort rules update ([#10167](https://github.com/SAP/fundamental-ngx/issues/10167)) ([457bfac](https://github.com/SAP/fundamental-ngx/commit/457bfac0dab855012614a02c802366c296479ab5))
+
+
+
+
+
 # [0.43.0-rc.7](https://github.com/SAP/fundamental-ngx/compare/v0.43.0-rc.6...v0.43.0-rc.7) (2023-07-12)
 
 **Note:** Version bump only for package @fundamental-ngx/platform